Use `h5netcdf` instead of `netCDF4`
The `h5netcdf` package seems to be better supported between python
versions and is a little bit lighter. It is used to read the netcdf
files directly instead of via `xarray` to read the original 2-d
"altitude" variable where `xarray` fails. However, `h5netcdf` does not
read netcdf3 files, so we fall back to `scipy.io` in those cases.
And we need to take care of the slightly different variable attribute
handling.