Return None if monthly l2 file cannot be found
Returns None for the dataset list containing the single species datasets
if one of the corresponding monthly files is not found.
This ensures that for all of the species required for the calculations,
we have a dataset. Whether that contains lots of useful data is another
matter and should be handled in a different place.