ci: Manually downgrade `pandas` for Py38
`xarray` changed its minimum python version to 3.9 in 2023.2:
https://github.com/pydata/xarray/pull/7461 but a "fix" (version cap)
for `pandas` was introduced in `xarray` version 2023.2, which will not
be installed on python 3.8 setups. Pip/conda will happily pull the
latest `pandas` version (now 2.x which is incompatible with `xarray`).
To remedy the situation, manually downgrades the `pandas` version for
the Py38 CI setup.