util: Remove separate first, rest file reading
We used to read the first data set (file) separately to interpolate the
altitudes of the rest to the ones from the first. This interpolation was
the only distinction and it is now handled elsewhere, so we can just
read the files all the same. No need to treat the first one special.