epp: Add NOy background subtraction for EPP
Includes the functions to subtract the NOy background based on
the prepared NOy/CH4 correlation analysis. Adjusts the subtraction based
on different criteria, e.g. CO and potential temperature.
The thresholds are taken from the IDL code.

Includes functions for processing one profile at a time as well as for
processing many at once. Both should produce the same results, but are
not fully tested yet.

Also provides a function to integrate (sum) the calculated NOy number of
molecules over a volume. Takes some care with the lower end to avoid
mixing too much non-EPP NOy in the sum.