epp: Don't limit max altitude when smoothing
The altitude range was limited for smoothing targets to the maximum
available from both. This was done mainly to avoid having NaN in the
resolution vector which leads to complications when convolving with the
Gaussian weights. Instead we extrapolate the resolution vector with
zeros so that the lower values get convolved in to produce profiles for
the whole altitude range.