epp: Select nearest points in multi-processing
Selects the nearest points in time and latitude to interpolate the
background. When not using the pre-calculated and daily-interpolated
table, some indices are not in the index and processing fails.
Using the nearest available point in those cases to be able to continue
with the processing. Leaves the issue of the lookup table being correct
to the caller.