epp: Enable iteration over regions for correlation
Iterates over regions when calculating tracer-tracer correlations.
Facilitates calculating northern and southern hemisphere together before
using that to subtract as a background. Makes it easier to combine
afterwards, and can be used in cases where a pre-calculated background
correlation data set is not (yet) available.

The plotting and total amount calculations are still for one hemisphere
only.