epp: Iterate over regions for the histogram plots
Iterates over the regions for plotting the histograms and appends the
region number to the file name. Enables us to do tracer-tracer
correlations for north and south in one run, having a histogram for each
if we want to. Otherwise it also fails because of the new latitude
dimension in the histogram data set.