epp: Save EPP-NOy data set
Saves the EPP-NOy data set in the same format as the "original"
MIPAS v8 data. This replaces the `target` variable and formats the file
as is done for the NOy data itself.
Based on the NOy data set, all other variables are the same as for NOy.
To be used to produce EPP-NOy as a level-2b data product.