util: Fix altitude dtype (mix) for pandas 2
Pandas 2 does not like big endian indices on little endian systems, and
probably neither the other way around. Pandas 1.5 converted the index
dtype implicitly, but version 2 does not seem to do that anymore, so big
endian stays big endian and index lookups fail.
Especially when reading files with different packages, xarray, h5netcdf,
or scipy.io, mixing dtypes is not a good idea.

Fixes the dtype of the altitude variable and coordinate to be the same
as the one of the `target` variable read with xarray to make things work
with pandas 1.5 and 2.