epp: Unset default integration CO threshold
Sets the default CO threshold for the hemispheric integration to None.
We check for it being `None` to decide between finding the minimum
NOy/CH4 ratio or using the CO threshold. Makes it consistent with being
unset in the configuration file, to mean use the minimum ratio.
There is actually no "None" in toml.