epp: Drop latitude from selected correlation data
Drops the `latitude` variable/coordinate from the per-hemisphere
selected correlation value before interpolating to the actual CH4.
The dataset/array for interpolation may carry its own coordinates which
may conflict with the selected one. This avoids such a conflict, and using
the coordinates of the data we interpolate to is what we are interested
in anyway. No need to keep the coordinates of the "original" NOy-CH4
correlation data set.