epp: Separate NH and SH when looking for min alt
Adapts the three (or `lsearch_num`) latitude bands for the SH and NH
when looking for the altitude of the minimum NOy/CH4.
Assumes that the latitudes are ordered from low (SH) to high (NH), so
that the first indices indicate the southern polar border, and the last
indices indicate the northern polar border of the data.