epp: Use correlation DataArray instead of Dataset
Leaves it to the caller which of the correlation statistics is used for
subtracting the background NOy. Makes it easier to switch from e.g.
"mean" to "median" or "mode". However the single profile version is a
bit trickier because we reset the coordinates after interpolation to
avoid any overlap with the main dataset.