Ntot: Use abs(grad) and set vertical axis
Uses the absolute value of the altitude gradient for the layer thickness
to avoid negative volumes. For example on pressure levels the altitude
may be swapped and the differences are negative.
Also for pressure-level interpolated data, the altitude may be two (or
more) dimensional, so we allow for an extra vertical axis argument to
pass through. Basically we only need the gradient (layer thickness) in
the vertical (altitude or pressure) direction.