epp: Add option to smooth vertical resolution
When combining the CH4 and NOy data, the CH4's resolution is degraded to
the one from NOy. In the IDL code, both resolution vectors are filtered
through a boxcar running average of 4 points before this convolution.
Adds an option to do the same here, but a little more flexible, the
length is the argument, so for IDL compatibility, pass `smooth_vr=4`.