epp cli: Use configured CH4 akd threshold
Uses the CH4 ak diagonal threshold to select valid data, if configured.
Skips the variables that have no `altitude` dimension to avoid
unnecessary broadcasting. This needs to be probably made a bit more
flexible, e.g. selecting variables with common dimensions.

Putting it into `correlation.thresholds` makes the most sense, but it is
not used later in `calc_noy_bg_epp()`, so we `pop` it from the
dictionary.