epp: Use `time` as the base dimension for datasets
In most of the data sets we kept the `geo_id` dimension as the basis for
processing and output because the strings are supposed to be unique.
Switches to `time` to be consistent with the "standard" MIPAS V8 netcdf,
in particular for the EPP-NOy output, so that people can use that in the
same way as the normal level-2 products.