epp: Use `int64` milli/nano-second time resolution
Rounds the read data sets' time variables to 1 ms, since the conversion
from double "days since ..." to `numpy`'s `datetime64[ns]` (= int64
nanoseconds since 1970-01-01 00:00) is not exact and the times can be
off a few microseconds. Rounding to 1 ms should be good enough.

When serializing to netcdf we can then use `int64` millisecond which are
then round-trip safe. Except for the histogram dataset where the time
variable is the average of the monthly data and needs more digits.
Instead of rounding them, we serialize with nanosecond resolution.