epp: Fix CH4 AKD threshold application
We `.pop()`ed the CH4 AKD threshold from the dictionary within the
(monthly) iteration loop, so this was only applied once for the first
month. However, this threshold should be applied to all months.
Moves the reading part outside of the loop so that it persists
throughout the whole iteration.