Refine weighted zonal mean weights handling
A bit inspired by xarray's handling, avoids the creation of a dummy
"weights" dataset variable for broadcasting and uses `.where()` to
promote the shape(s) to the dataset variable shapes.
Using `.where()` and setting the units explicitly has the advantage that
it seems to work with unitful arrays too, as the addition of scalars to
them be a bit fragile.