epp: Use star-import for unit handling
We relied on `to_unit()` to be provided by the local "units" module.
That was just cut short to use a separate module for that, but we need
to ensure that units are available with xarray, so we just import
everything. Might not be needed after all, but keep it in for now.