epp: Guard against non-existing name attributes
When calculating densities and molecules, we update the `long_name` and
`standard_name` attributes by replacing strings. This can fail when the
attributes do not exist and we try to access them via ["..."].
Uses `.get()` and an almost empty default string to set these attributes
to something at least a little sensible.