epp: Swap number and array to avoid conversion
Swaps the number and array to avoid any unintended conversion of units.
The minimum index from `.argmin()` and seems to still carry units,
usually "ppm". Adding 1 to it would get converted to 1e6 ppm, but we
just need the index. Swapping the number and the array avoids that kind
of conversion resulting in the correct altitude index.