epp: Expand ESE array for multiple pressures
Adapts the shape of the result array in cases where the pressure(s) are
passed as a list or array. `dn` should be a single-element array to get
a scalar for `zeros()`, we take the first element only. The correct
shape of `dn` (and hence `t`) is currently not enforced.